requirements.txt
anyio>=3.4.0
asgiref>=3.4.1
azure-core>=1.21.1
azure-storage-blob>=12.9.0
certifi>=2021.10.8
cffi>=1.15.0
charset-normalizer>=2.0.9
scikit-image
click>=8.0.3
colorama>=0.4.4
cryptography>=36.0.1
cycler>=0.11.0
dnspython>=2.1.0
email-validator>=1.1.3
fastapi>=0.70.1
wheel
# fonttools>=4.28.5
h11>=0.12.0
httptools>=0.2.0
idna>=3.3
imageio>=2.13.5
isodate>=0.6.1
itsdangerous>=2.0.1
Jinja2>=3.0.3
joblib>=1.1.0
kiwisolver>=1.3
MarkupSafe>=2.0
matplotlib>=3
msrest>=0.6
networkx>=2
numpy>=1.2
oauthlib>=3.1
orjson>=3.6
packaging>=21.3
Pillow>=8.4
pycparser>=2.21
pydantic>=1.8
pyparsing>=3.0
python-dateutil>=2.8
python-dotenv>=0.19
python-multipart>=0.0.5
PyWavelets>=1
PyYAML>=5
requests==2.27
requests-oauthlib>=1.3
scikit-learn>=0.24.2
# scipy
six>=1.11
sniffio<=1.2
starlette<=0.16
threadpoolctl<=3
tifffile<=2021.11
typing-extensions<=4
# ujson<=4
urllib3<=1.27
uvicorn<=0.15
watchgod<=0.7
websockets<=10
pytest
